requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
Flask-Login
python-dotenv
Flask-WTF
Flask-Bcrypt
email_validator